Migrate data_table_manager to Galaxy.
